requirements.txt
lxml==4.1.0
python-docx==0.8.6
xlrd==1.1.0
